+++ packages/core/src/components/AppLoader/index.tsx
@@ -4,6 +4,7 @@ import ErrorMessage from '../ErrorMessage';
 import Loader from '../Loader';
 import Title from '../Title';
 import Text from '../Text';
+import { ErrorType } from '../../types';
 
 export type Props = {
   /** Center the loader and content. */
@@ -11,7 +12,9 @@ export type Props = {
   /** Content to be rendered on a successful request. */
   children: NonNullable<React.ReactNode>;
   /** Request error. */
-  error?: Error | null;
+  error?: ErrorType | null;
+  /** Title to display on error message. */
+  errorTitle?: React.ReactNode;
   /** Text to display on a failed request. */
   failureText: NonNullable<React.ReactNode>;
   /** Whether a request is fetched. */
@@ -37,6 +40,7 @@ export class AppLoader extends React.Component<Props & WithStylesProps> {
       centered,
       children,
       error,
+      errorTitle,
       failureText,
       fetched,
       loadingText,
@@ -60,7 +64,7 @@ export class AppLoader extends React.Component<Props & WithStylesProps> {
         )}
 
         <div {...css(styles.errorOrLoader)}>
-          {error ? <ErrorMessage error={error} /> : <Loader inline />}
+          {error ? <ErrorMessage error={error} title={errorTitle} /> : <Loader inline />}
         </div>
       </div>
     );