+++ packages/metrics/src/utils/hasGoogleAnalytics.ts
@@ -1,0 +1,3 @@
+export default function hasGoogleAnalytics() {
+  return typeof ga === 'function';
+}